Disable -fPIE during Dropbear hardening
Since we know we're producing a .so, adding -fPIE is incompatible with -fPIC
which we'll be using on amd64 builds. The -fPIE option is added in order to
harden Dropbear and there's no option that will disable it besides making
a static build, which we are not aiming for. So we just disable the entire
block in configure with an `if false` so pie will not be included. Mmmmmm, pie.